+++ core/src/com/mygdx/game/Enemies/DefendedCannon.java
@@ -11,8 +11,10 @@ import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.physics.box2d.PolygonShape;
 import com.badlogic.gdx.physics.box2d.World;
 import com.badlogic.gdx.utils.Array;
+import com.mygdx.game.Graphics.Assets;
 import com.mygdx.game.Levels.Level1.GameScreenLevel1;
 import com.mygdx.game.Objects.Bullet;
+import com.mygdx.game.Graphics.HUD;
 
 import static com.mygdx.game.MyGame.*;
 
@@ -26,7 +28,7 @@ public class DefendedCannon extends Enemy {
     private boolean setToDestroy;
     private boolean destroyed;
     private boolean attack;
-    private static final float shootTimer = 2f;
+    private static final float shootTimer = 1f;
     private float stateTimer;
     private float timer;
     private Animation BTOOM;
@@ -35,8 +37,8 @@ public class DefendedCannon extends Enemy {
     private Sprite btoom3;
     public Array<Bullet> defendedCannonBullets = new Array<Bullet>();
 
-    public DefendedCannon(World world, float x, float y) {
-        super(world, x, y);
+    public DefendedCannon(World world, float x, float y, boolean flip) {
+        super(world, x, y, flip);
         HP = 10;
         atlas = new TextureAtlas("Animations/Btoom.txt");
         btoom1 = atlas.createSprite("Btoom1");
@@ -46,9 +48,12 @@ public class DefendedCannon extends Enemy {
         frames.add(btoom1);
         frames.add(btoom2);
         frames.add(btoom3);
-        BTOOM = new Animation(0.1f, frames);
+        BTOOM = new Animation(0.05f, frames);
         textureCannon = new Texture("Enemies/DefendedCannon.png");
         spriteCannon = new Sprite(textureCannon);
+        if(isFlip){
+            spriteCannon.flip(true, false);
+        }
         setBounds(getX(), getY(), 32/PPM, 32/PPM);
         setToDestroy = false;
         destroyed = false;
@@ -57,21 +62,18 @@ public class DefendedCannon extends Enemy {
     public void update(float delta){
         stateTimer += delta;
         timer += delta;
-        if(setToDestroy && !destroyed){
+        if(setToDestroy && !destroyed && stateTimer > 0.3f){
             world.destroyBody(b2body);
             attack = false;
             setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
             setRegion((Sprite)BTOOM.getKeyFrame(timer, false));
             stateTimer = 0;
             destroyed = true;
-        }
-        else if(!destroyed) {
-            setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
-            setRegion(spriteCannon);
+            HUD.SCORE+=35;
         }
         if(attack){
             if (stateTimer >= shootTimer) {
-                if (!spriteCannon.isFlipX()) {
+                if (!isFlip) {
                     bullet = new Bullet(world, b2body.getPosition().x, b2body.getPosition().y, 24/PPM, 1/PPM);
                     bullet.bulletBody.setLinearVelocity(4f, 0);
                     defendedCannonBullets.add(bullet);
@@ -84,16 +86,13 @@ public class DefendedCannon extends Enemy {
             }
         }
         for (Bullet bullet:defendedCannonBullets){
-            bullet.update(delta);
+            bullet.update();
             if(bullet.isDestroyed())
                 playerBullets.removeValue(bullet, true);
         }
     }
 
     public void draw(Batch batch){
-        if(stateTimer == 0 || !destroyed){
-            super.draw(batch);
-        }
         for (Bullet bullet:defendedCannonBullets){
             if(!bullet.isDestroyed()) {
                 bullet.draw(batch);
@@ -131,10 +130,18 @@ public class DefendedCannon extends Enemy {
 
         PolygonShape back = new PolygonShape();
         Vector2[] weakPoint = new Vector2[4];
-        weakPoint[0] = new Vector2(-14, 16).scl(1/PPM);
-        weakPoint[1] = new Vector2(-14, -16).scl(1/PPM);
-        weakPoint[2] = new Vector2(-13, -16).scl(1/PPM);
-        weakPoint[3] = new Vector2(-13, 16).scl(1/PPM);
+        if(!isFlip) {
+            weakPoint[0] = new Vector2(-14, 16).scl(1 / PPM);
+            weakPoint[1] = new Vector2(-14, -16).scl(1 / PPM);
+            weakPoint[2] = new Vector2(-13, -16).scl(1 / PPM);
+            weakPoint[3] = new Vector2(-13, 16).scl(1 / PPM);
+        }
+        else {
+            weakPoint[0] = new Vector2(14, 16).scl(1 / PPM);
+            weakPoint[1] = new Vector2(14, -16).scl(1 / PPM);
+            weakPoint[2] = new Vector2(13, -16).scl(1 / PPM);
+            weakPoint[3] = new Vector2(13, 16).scl(1 / PPM);
+        }
         back.set(weakPoint);
 
         fdef.shape = back;
@@ -167,8 +174,29 @@ public class DefendedCannon extends Enemy {
     @Override
     public void deleted() {
         setToDestroy = true;
+        stateTimer = 0;
     }
+
+    @Override
+    public void reverseVelocity(boolean x, boolean y) {
+
+    }
+    @Override
     public boolean isDestroyed(){
         return destroyed;
+    }
+
+    public Sprite babax(){
+        Sprite sprite;
+        if(setToDestroy && !destroyed){
+            sprite = (Sprite) BTOOM.getKeyFrame(stateTimer, false);
+        }
+        else if (destroyed){
+            sprite = Assets.spriteEmpty;
+        }
+        else {
+            sprite = spriteCannon;
+        }
+        return sprite;
     }
 }
